Commit Title: Add module import function to main interface
The main interface now has a new button for importing modules from a specific program of study. This functionality is limited to the field of Informatics B.Sc. The code for the classes MainUI.java, Main.java, Database.java and Studiengang.java were modified, also the file MainUI.fxml.

The new function was added to improve usability, avoiding manual input for known modules and making the application more dynamic and easier to use.